pyproject.toml
[tool.poetry]
name = "frank"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pymongo = "^3.11.0"
dnspython = "^2.0.0"
getkey = "^0.6.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
